02_Dense_Vector_Retrieval_V2/pyproject.toml
[project]
name = "03-enhanced-rag-metadata"
version = "2.0.0"
description = "Enhanced RAG with Metadata Filtering and Multiple Distance Metrics"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel>=6.29.5",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.0",
    "nest-asyncio>=1.6.0",
    "numpy>=2.0.0",
    "openai>=1.59.7",
    "pandas>=2.2.3",
    "plotly>=5.24.1",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.1",
]
